storage: extract ReportSSTEntries
It's going to come in handy for testing in subsequent commits.
